Fix regressions from namespacing Update template locations Add permission codes to cms sections Code, phpdoc, and typehint cleanup Split batch actions into files
