+++ _config/legacy.yml
@@ -3,4 +3,4 @@ Name: cmslegacy
 ---
 SilverStripe\ORM\DatabaseAdmin:
   classname_value_remapping:
-   SiteTree: 'SilverStripe\CMS\Model\SiteTree'
+    SiteTree: 'SilverStripe\CMS\Model\SiteTree'
