+++ code/BatchActions/CMSBatchAction_Publish.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace SilverStripe\CMS\BatchActions;
+
+use SilverStripe\Admin\CMSBatchAction;
+use SilverStripe\ORM\SS_List;
+
+/**
+ * Publish items batch action.
+ *
+ * @package cms
+ * @subpackage batchaction
+ */
+class CMSBatchAction_Publish extends CMSBatchAction
+{
+	public function getActionTitle()
+	{
+		return _t('CMSBatchActions.PUBLISH_PAGES', 'Publish');
+	}
+
+	public function run(SS_List $pages)
+	{
+		return $this->batchaction($pages, 'publishRecursive',
+			_t('CMSBatchActions.PUBLISHED_PAGES', 'Published %d pages, %d failures')
+		);
+	}
+
+	public function applicablePages($ids)
+	{
+		return $this->applicablePagesHelper($ids, 'canPublish', true, false);
+	}
+}
