+++ code/Controllers/ContentController.php
@@ -19,7 +19,6 @@ use Controller;
 use Page;
 
 use SiteConfig;
-use Config;
 use SS_HTTPRequest;
 use Translatable;
 use i18n;
