+++ templates/SilverStripe/CMS/Controllers/Includes/AssetAdmin_UploadContent.ss
