+++ templates/SilverStripe/CMS/Controllers/Includes/CMSMain_Tools.ss
