+++ templates/SilverStripe/CMS/Controllers/Includes/CMSPageAddController_Content.ss
