+++ templates/SilverStripe/CMS/Controllers/Includes/CMSPageHistoryController_Tools.ss
