+++ templates/SilverStripe/CMS/Controllers/Includes/CMSPagesController_Tools.ss
